rename pgp.js to openpgpjs-custom.js
