+++ extension/js/common/core/crypto/pgp/pgp-key.ts
@@ -6,7 +6,7 @@ import { Buf } from '../../buf.js';
 import { Catch, UnreportableError } from '../../../platform/catch.js';
 import { MsgBlockParser } from '../../msg-block-parser.js';
 import { PgpArmor } from './pgp-armor.js';
-import { opgp } from './pgp.js';
+import { opgp } from './openpgpjs-custom.js';
 import { OpenPGPKey } from './openpgp-key.js';
 import { SmimeKey } from '../smime/smime-key.js';
