+++ extension/js/common/core/crypto/pgp/pgp-msg.ts
@@ -7,7 +7,7 @@ import { Value } from '../../common.js';
 import { Buf } from '../../buf.js';
 import { Catch } from '../../../platform/catch.js';
 import { PgpArmor, PreparedForDecrypt } from './pgp-armor.js';
-import { opgp } from './pgp.js';
+import { opgp } from './openpgpjs-custom.js';
 import { KeyCache } from '../../../platform/key-cache.js';
 import { ContactStore } from '../../../platform/store/contact-store.js';
 import { SmimeKey } from '../smime/smime-key.js';
