+++ extension/js/common/platform/store/contact-store.ts
@@ -3,7 +3,7 @@
 import { PgpClient } from '../../api/pub-lookup.js';
 import { AbstractStore } from './abstract-store.js';
 import { Catch } from '../catch.js';
-import { opgp } from '../../core/crypto/pgp/pgp.js';
+import { opgp } from '../../core/crypto/pgp/openpgpjs-custom.js';
 import { BrowserMsg } from '../../browser/browser-msg.js';
 import { Str } from '../../core/common.js';
 import { PgpKey, Pubkey, Contact } from '../../core/crypto/pgp/pgp-key.js';
